src/main/java/br/com/alura/tdd/service/BonusService.java
package br.com.alura.tdd.service;

import java.math.BigDecimal;
import java.math.RoundingMode;

import br.com.alura.tdd.modelo.Employee;
import br.com.alura.tdd.modelo.Funcionario;

public class BonusService {

	public BigDecimal calcularBonus(Employee employee) {
		BigDecimal valor = employee.getSalario().multiply(new BigDecimal("0.1"));
		if (valor.compareTo(new BigDecimal("1000")) > 0) {
			throw new IllegalArgumentException("Funcionário com salário maior do que R$ 10000. Não pode receber bônus.");
		}
		return valor.setScale(2, RoundingMode.HALF_UP);
	}

}
